requirements.txt
certifi==2018.4.16
chardet==3.0.4
click==6.7
decorator==4.3.0
Flask==1.0.2
gunicorn==19.8.1
idna==2.6
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
MarkupSafe==1.1.1
nbformat==4.4.0
numpy==1.14.3
pandas==0.22.0
pipenv==11.10.4
plotly==2.6.0
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
six==1.11.0
traitlets==4.3.2
urllib3==1.22
virtualenv==15.2.0
virtualenv-clone==0.3.0
Werkzeug==0.14.1
tensorflow==1.3.0
tensorflow-tensorboard==0.1.8
scikit-learn==0.19.1
requests==2.18.4
Keras==2.0.9
glob2==0.5
Pillow==4.2.1